Update uv sync command in README.md
As far as I can tell, running `uv sync` by itself removes the install of the package itself because it's not listed in the `requirements-dev.txt`.  This PR tacks ` && python -m pip install -e .` on to the end of the `uv sync` command so that when new dev requirements are added to the project you can easily copy/paste a command that will sync the requirements and still end up with the package itself installed.